feat: useOpenResolutionModal 커스텀훅으로 빼기
