+++ src/components/common/Banner/WelcomeBanner/index.tsx
@@ -8,6 +8,7 @@ import useAlert from '@/components/common/Modal/useAlert';
 import useModalState from '@/components/common/Modal/useModalState';
 import Responsive from '@/components/common/Responsive';
 import ResolutionModal from '@/components/resolution/ResolutionModal';
+import { useOpenResolutionModal } from '@/components/resolution/useOpenResolutionModal';
 import desktop34Banner1 from '@/public/icons/img/banner_34_desktop_ver1.gif';
 import desktop34Banner2 from '@/public/icons/img/banner_34_desktop_ver2.gif';
 import mobile34Banner1 from '@/public/icons/img/banner_34_mobile_ver1.gif';
@@ -56,31 +57,8 @@ const WelcomeBanner = ({ is34 }: WelcomeBannerProp) => {
     mobile: { 1: mobileOthersBanner1.src, 2: mobileOthersBanner2.src },
   };
 
-  const {
-    isOpen: isOpenResolutionModal,
-    onOpen: onOpenResolutionModal,
-    onClose: onCloseResolutionModal,
-  } = useModalState();
-
-  const { alert } = useAlert();
-  const { data } = useGetResolutionValidation();
-
-  const handleResolutionModalOpen = () => {
-    if (!data?.isRegistration) {
-      alert({
-        title: '편지는 한번만 전송할 수 있어요',
-        description: '전송된 편지는 종무식 때 열어볼 수 있어요!',
-        buttonText: '확인',
-        maxWidth: 400,
-        zIndex: zIndex.헤더 + 101,
-        buttonColor: colors.white,
-        buttonTextColor: colors.black,
-        hideCloseButton: true,
-      });
-    } else {
-      onOpenResolutionModal();
-    }
-  };
+  const { isOpenResolutionModal, onCloseResolutionModal, handleResolutionModalOpen, memberProfileImgUrl } =
+    useOpenResolutionModal();
 
   return (
     <WelcomeBannerContainer>
@@ -94,10 +72,7 @@ const WelcomeBanner = ({ is34 }: WelcomeBannerProp) => {
                     NOW, 다짐하러 가기
                   </ResolutionButton>
                   {isOpenResolutionModal && (
-                    <ResolutionModal
-                      profileImageUrl={data?.memberProfileImgUrl ?? ''}
-                      onClose={onCloseResolutionModal}
-                    />
+                    <ResolutionModal profileImageUrl={memberProfileImgUrl ?? ''} onClose={onCloseResolutionModal} />
                   )}
                 </ButtonWrapper>
                 <BannerWrapper>
